chore(meshwide): split sync to node logic
